Move GraphDataFetch to production code
