+++ street/src/test-fixtures/java/org/opentripplanner/street/graph/DisposableEdgeDataFetcher.java
@@ -3,7 +3,6 @@ package org.opentripplanner.street.graph;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.opentripplanner.street.GeoJsonIo;
 import org.opentripplanner.street.linking.DisposableEdgeCollection;
 
 public class DisposableEdgeDataFetcher {
